[AMS] IcebergPartitionPlan impl buildTaskSplitter by Bin-Packing
