+++ ams/server/src/main/java/com/netease/arctic/server/optimizing/plan/IcebergPartitionPlan.java
@@ -23,8 +23,6 @@ import com.netease.arctic.optimizing.OptimizingInputProperties;
 import com.netease.arctic.server.table.TableRuntime;
 import com.netease.arctic.table.ArcticTable;
 
-import java.util.Collections;
-
 public class IcebergPartitionPlan extends AbstractPartitionPlan {
 
   protected IcebergPartitionPlan(TableRuntime tableRuntime, ArcticTable table, String partition, long planTime) {
@@ -33,8 +31,7 @@ public class IcebergPartitionPlan extends AbstractPartitionPlan {
 
   @Override
   protected TaskSplitter buildTaskSplitter() {
-    // TODO not split tasks in a partition now
-    return targetTaskCount -> Collections.singletonList(new SplitTask(fragmentFiles, segmentFiles));
+    return new BinPackingTaskSplitter();
   }
 
   @Override
